requirements.txt
opencv-contrib-python
pyrebase4
face-library